Updated acount.js and account.js with working functionality.
