--- account.ejs
@@ -1,71 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Account Page</title>
-    <link href="https://fonts.googleapis.com/css2?family=Roboto:wght@400;500&display=swap" rel="stylesheet">
-    <link rel="stylesheet" href="/stylesheets/account-style.css">
-    <link rel="stylesheet" href="/stylesheets/navbar-style.css">
-</head>
-<body>
-    <div class="navbar">
-        <a href="/">Home</a>
-        <a href="/create-listing">Create Listing</a>
-        <a href="/messages">Messages</a>
-        <div class="navbar-right">
-            <a href="/help">Help</a>
-            <a class="active" href="/account">Account</a>
-            <a href="/signup">Sign up</a>
-            <a href="/login">Log in</a>
-        </div>
-    </div>
-    <div class="container">
-        <div class="box">
-            <div class="box-content">
-                <h2 class="header">Account Details</h2>
-                <form action="/account/update" method="POST" class="account-form">
-                    <label for="name">Name:</label>
-                    <input type="text" id="name" name="name" value="<%= user.name %>" required>
-
-                    <label for="username">Username:</label>
-                    <input type="text" id="username" name="username" value="<%= user.username %>" required>
-
-                    <label for="email">Email:</label>
-                    <input type="email" id="email" name="email" value="<%= user.email %>" required>
-
-                    <label for="password">Password:</label>
-                    <input type="password" id="password" name="password" placeholder="Enter new password">
-
-                    <button type="submit">Update Account</button>
-                </form>
-            </div>
-        </div>
-
-        <div class="box">
-            <div class="box-content">
-                <h2 class="header">Active Listings</h2>
-                <ul class="list3">
-                    <li>No Items Currently Listed</li>
-                </ul>
-            </div>
-        </div>
-        <div class="box">
-            <div class="box-content">
-                <h2 class="header">Past Purchases</h2>
-                <ul class="list2">
-                    <li>
-                        <div class="item-box">Listing 1<br> Bought Feb 22 2025</div>
-                    </li>
-                    <li>
-                        <div class="item-box">Listing 2<br> Bought Jan 30 2025</div>
-                    </li>
-                    <li>
-                        <div class="item-box">Listing 3<br> Bought Dec 13 2024</div>
-                    </li>
-                </ul>
-            </div>
-        </div>
-    </div>
-</body>
-</html>
